HumanClassifier now optimizes its internal parameters
The basic functionalities of both HumanClassifier and HumanRegressor are now working as intended. I would really like to make this available as a pip package, and potentially as a scikit-learn compatible project, but from a preliminary analysis, I will probably need to move most of the coherence checks on input from method __init__ to fit, and I am not sure how to manage get_params and set_params. To be further investigated.